making code more modular and creating another css file to split up the amount of code. Created another file within dashboard for other links
