+++ src/Components/App.js
@@ -5,10 +5,10 @@ import { BrowserRouter as Router, Route,Switch} from 'react-router-dom';
 import Home from "./Login/Home"
 import LandingPage from "../Components/LandingPage"
 import Dashboard from "./Dashboard/Dashboard"
-import Inspiration from "./Dashboard/Inspiration"
-import RecipeMaker from "./Dashboard/RecipeMaker"
-import Social from "./Dashboard/Social"
-import UserAccount from "./Dashboard/UserAccount"
+import Inspiration from "./Dashboard/LinksWithinDash/Inspiration"
+import RecipeMaker from "./Dashboard/LinksWithinDash/RecipeMaker"
+import Social from "./Dashboard/LinksWithinDash/Social"
+import UserAccount from "./Dashboard/LinksWithinDash/UserAccount"
 import Register from "./Login/Register"
 
 export default function App() {
