+++ src/Components/Dashboard/LinksWithinDash/EditUserAccount.js
@@ -1,6 +1,6 @@
 import React from "react"
 import Button from "@material-ui/core/Button"
-import "../../styles.css"
+import "./linksWithinDash.css"
