+++ src/Components/Dashboard/LinksWithinDash/linksWithinDash.css
@@ -1,0 +1,99 @@
+/* INSPERATIONAL RECIPES */
+
+.inspoRecipe{
+    border: 2px solid black;
+    width: 30%;
+    margin: 0.5%;
+    border-radius: 3px;
+    box-shadow: 3px 3px 3px rgb(167, 163, 163)
+  }
+  .inspoRecipe p{
+    margin: 5%;
+  }
+  .inspoRecipe img{
+    width: 100%;
+  }
+  .inspoRecipeContainer{
+    display: flex;
+    flex-wrap: wrap;
+  }
+
+
+/*EDIT AND DELETE RECIPE */
+.editRecipeDisplay{
+    border: 1px solid black;
+    display: flex;
+    flex-wrap: wrap;
+    width: 50%;
+    margin: 0 auto;
+    margin-top: 8%;
+  }
+  
+  .editRecipeInput{
+    border: 1px solid grey;
+    width:60%;
+    font-size: 16px;
+  
+  }
+  .editRecipeDisplayInput{
+    width: 100%;
+    display: flex;
+    margin: 2% 0 3% 0
+  }
+  
+  .editRecipeDisplayInput p{
+    margin-left: 1%;
+    margin-right: 4%;
+  }
+  .editRecipeButtons{
+    width: 100%;
+    display: flex;
+    justify-content: space-between;
+    margin-top: 10%;
+    margin-bottom: 2%;
+    padding:2%;
+  }
+  .editRecipeDisplayInput:first-child input{
+    margin-left: 2.5%;
+  }
+  .editRecipeInput.editRecipeLastInput{
+    margin-left: -0.5%;
+  }
+
+
+  /* RECIPE MAKER PAGE */
+.recipeForm{
+    width: 30%;
+    margin: 2% auto auto auto;
+    border: 1px solid black;
+    background-color: #F1F0F0;
+    border-radius: 3px;
+    padding:3%;
+  }
+  .recipeListDisplay{
+    border:1px solid black;
+    width:30%;
+    margin: 0 auto;
+    margin-top: 3%;
+    border-radius: 3px;
+    box-shadow: 3px 3px 3px rgb(167, 163, 163)
+  }
+  .successfullyCreate{
+    margin: 0 auto;
+    width: 40%;
+    margin-top: 2%;
+  }
+  .vanish{
+    display: none
+  }
+
+
+
+/* SOCIAL PAGE */
+
+.user{
+    border:1px solid red;
+    margin:2% auto;
+    width: 80%;
+    padding:3%;
+  }
