+++ src/styles.css
@@ -23,6 +23,10 @@
   height:8vh;
   opacity: 0.5; */
 }
+.landingPageButtons{
+  display: flex;
+  justify-content: space-around;
+}
 
 .landingPageName{
   position: absolute;
@@ -46,22 +50,33 @@
   background-color: rgb(255, 255, 255);
   padding-top: 2%;
 }
+.registerRedirect{
+  color:blue;
+  cursor: pointer;
+}
+.registerPage{
+  margin-top: 13.4%;
+}
 .registerBoxText{
   font-weight: 900;
   padding-bottom: 20%;
 }
-.landingPageButtons{
-  display: flex;
-  justify-content: space-around;
-}
+
 
 /* DASHBOARD */
 
+.CardTitle{
+  color:white;
+  background: #008BF1;
+  font-family: 'Roboto', sans-serif;
+  padding: 10px;
+  border-top-left-radius: 3%;
+  border-top-right-radius: 3%;
+}
 .dashboard{
   border: 1px solid black;
   height: 100rem
 }
-
 .dashboardMain{
   display:flex;
   justify-content: space-between;
@@ -71,7 +86,6 @@
   display: flex;
 
 }
-
 .DashboardBlock{
   background-color: rgb(241, 240, 240);
   margin: 0 auto;
@@ -81,32 +95,20 @@
   font-family: 'Roboto', sans-serif;
   width: 15%;
 }
-.CardTitle{
-  color:white;
-  background: #008BF1;
-  font-family: 'Roboto', sans-serif;
-  padding: 10px;
-  border-top-left-radius: 3%;
-  border-top-right-radius: 3%;
-}
-
 .dashboardComp{
   width:99%;
 
 }
-
-
-
 .dashboardHeader{
   border:1px solid black;
 }
+
 
 /* HOME */
 .loginScreen{
   margin: 8% auto;
   padding-top:10%;
 }
-
 .loginBox{
   border:1px solid black;
   width: 30%;
@@ -121,22 +123,6 @@
   font-size: 10%;
   border:1px solid black;
 }
-.registerRedirect{
-  color:blue;
-  cursor: pointer;
-}
-.registerPage{
-  margin-top: 13.4%;
-}
-/* .registerBox{
-  border:1px solid black;
-  width: 30%;
-  height: 19vh;
-  margin: 0 auto;
-  padding: 1%;
-  border-radius: 2%;
-  box-shadow: 3px 3px 3px rgb(167, 163, 163)
-} */
 .loginError{
   border: 1px solid #F44335;
   background-color:#F44335;
@@ -147,7 +133,6 @@
   box-shadow: 3px 3px 3px #DBDBDB;
   padding:1%;
 }
-
 
 /* LINKS */
 .link{
@@ -197,107 +182,10 @@
   justify-content: space-around;
   font-size: 25px;
 }
-
-/* RECIPE MAKER PAGE */
-.recipeForm{
-  width: 30%;
-  margin: 2% auto auto auto;
-  border: 1px solid black;
-  background-color: #F1F0F0;
-  border-radius: 3px;
-  padding:3%;
-}
-.recipeListDisplay{
-  border:1px solid black;
-  width:30%;
-  margin: 0 auto;
-  margin-top: 3%;
-  border-radius: 3px;
-  box-shadow: 3px 3px 3px rgb(167, 163, 163)
-}
-.successfullyCreate{
-  margin: 0 auto;
-  width: 40%;
-  margin-top: 2%;
-}
 .searchBar{
   border:1px solid black;
   font-size: 18px;
   border-radius: 10px;
   padding: 3%;
 }
-.vanish{
-  display: none
-}
 
-
-/* Edit recipe and delete */
-.editRecipeDisplay{
-  border: 1px solid black;
-  display: flex;
-  flex-wrap: wrap;
-  width: 50%;
-  margin: 0 auto;
-  margin-top: 8%;
-}
-
-.editRecipeInput{
-  border: 1px solid grey;
-  width:60%;
-  font-size: 16px;
-
-}
-.editRecipeDisplayInput{
-  width: 100%;
-  display: flex;
-  margin: 2% 0 3% 0
-}
-
-.editRecipeDisplayInput p{
-  margin-left: 1%;
-  margin-right: 4%;
-}
-.editRecipeButtons{
-  width: 100%;
-  display: flex;
-  justify-content: space-between;
-  margin-top: 10%;
-  margin-bottom: 2%;
-  padding:2%;
-}
-.editRecipeDisplayInput:first-child input{
-  margin-left: 2.5%;
-}
-.editRecipeInput.editRecipeLastInput{
-  margin-left: -0.5%;
-}
-
-/* INSPERATIONAL RECIPES */
-
-.inspoRecipe{
-  border: 2px solid black;
-  width: 30%;
-  margin: 0.5%;
-  border-radius: 3px;
-  box-shadow: 3px 3px 3px rgb(167, 163, 163)
-}
-.inspoRecipe p{
-  margin: 5%;
-}
-.inspoRecipe img{
-  width: 100%;
-}
-.inspoRecipeContainer{
-  display: flex;
-  flex-wrap: wrap;
-}
-
-
-/* SOCIAL PAGE */
-
-.user{
-  border:1px solid red;
-  margin:2% auto;
-  width: 80%;
-  padding:3%;
-}
